feat(AD/TFS): Support reviewers in PRs
You can now specify reviewers to be added to the pull requests created
for the Azure Devops (Server), and TFS or VSTS platforms.

You can do this by either specifying the `--reviewer` option on the
commandline, or by adding a `Reviewers` property to the
`nukeeper.settings.json` file.

As usual, CLI arguments take precedence over settings from the file
system.

The reviewer value should be an email address.

The current implementation uses the REST apis for teams, and team
members. However, I noticed there's also a users api that allows
querying based on the email address. I did not however see an easy way
to make sure that a particular user is part of a team by using that
API, as the `membersOf` property was empty in my case.

Somebody who is more familiar with these apis could probably improve
this implementation significantly but it is working for my azure devops
server instance.