fix(cli): Exit with <> 0 in case of failure
This commit is relevant for `CollaborationPlatformCommand`.

In many locations exceptions were being swallowed inappropriately. This
resulted in nukeeper running for a very long time as its current
behavior is to not stop until it can apply 1 update.

I've ensured that each separate repository gets a chance to run in case
of a multiple repository update command; single repository commands on
the other hand will now fail as soon as one of its update sets cannot be
applied.

The reasoning for this is the following: an error at the level of an
update set being applied is most likely going to happen for all the
different update sets. It is after all working with the same git
repository, target repository,... For example, if a push fails due to an
authentication issue, it will happen for all update sets.

After all repositories have had their chance and at least one of them
fails, an exception is throw that will result in a non-zero status code
being returned to process that invoked nukeeper.

`ICollaborationEngine` returns an integer, but this represents the
number of updates it has performed, and since there's no easy way to know
what the expected number of updates is at this point, there's no way to
utilize this information to be able to tell whether
`CollaborationPlatformCommand` should return a non-zero status code.