fix(nuget): Apply DependencyVersion Lowest
The default behavior of NuGet is to try to install the highest available
version for dependencies of a package. For some reason, it doesn't take
into account the target framework of these dependencies.

This means that, for example, for FluentValidation.WebApi 7.6.105, it
will try to install FluentValidation 9.5.1 as one of its dependencies;
FluentValidation 9.5.1 doesn't support .NET framework 4.5.2 as a target
framework.

A quick fix is to apply `DependencyVersion Lowest` instead.

This ensures that if only a single package is chosen to be updated, it
will at least still correctly update its dependencies to the minimal
required versions compared to the `Ignore` strategy.

Since packages aren't required to follow semantic versioning, this
should also be safer than the options `HighestPatch` and `HighestMinor`.

In long term, nukeeper should probably analyze the target framework of
its chosen updates, and ensure it matches with the target framework of
the projects it tries to update.

I will also raise an issue with NuGet to see if this is expected
behavior or not.